+++ src/App.js
@@ -2,11 +2,10 @@ import React, { useState } from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 // import $ from 'jquery';
 // import Popper from 'popper.js';
-import 'bootstrap/dist/js/bootstrap.bundle.min';
-import './components/atoms/main.scss';
+import './sass/main.scss';
 import AuthContext from './auth/context';
-import Login from './pages/Login';
 import Example from './examples/examples';
+import Login from './pages/login';
 
 function App() {
   const [user, setUser] = useState();
