+++ src/components/atoms/button/index.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import './style.scss';
+
+const Button = ({ title, ...rest }) => (
+  <button className='main-button' {...rest}>
+    {title}
+  </button>
+);
+
+export default Button;
