+++ src/components/atoms/button/style.scss
@@ -1,0 +1,13 @@
+.main-button {
+  border-radius: 4px;
+  background-color: black;
+  border: none;
+  color: white;
+  height: 50px;
+  width: 100%;
+  margin: 15px 0;
+  font-size: 17px;
+  &:hover {
+    opacity: 0.8;
+  }
+}
