+++ src/components/atoms/errorMessage/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import './style.scss';
+
+const ErrorMessage = ({ message }) => (
+  <div className='alert-danger'>{message}</div>
+);
+
+export default ErrorMessage;
