+++ src/components/atoms/errorMessage/style.scss
@@ -1,0 +1,5 @@
+.alert-danger {
+  color: red;
+  margin: 5px 0;
+  font-size: 16px;
+}
