--- src/components/atoms/form/Form.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-const Form = styled.form`
-  margin-top: 30px;
-`;
-
-export default Form;
