--- src/components/atoms/form/FormGroup.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-const FormGroup = styled.div`
-  margin: 5px 0;
-`;
-
-export default FormGroup;
