--- src/components/atoms/form/Label.js
@@ -1,16 +1,0 @@
-import colors from '../../base/colors';
-import styled from 'styled-components';
-
-const Label = styled.label`
-  width: 100%;
-  height: 14px;
-  font-size: 12px;
-  font-weight: 500;
-  font-stretch: normal;
-  font-style: normal;
-  line-height: normal;
-  letter-spacing: 1.13px;
-  color: ${colors.secondaryGray};
-`;
-
-export default Label;
