--- src/components/atoms/form/index.js
@@ -1,6 +1,0 @@
-export { default as Button } from './Button';
-export { default as ErrorMessage } from './ErrorMessage';
-export { default as FormGroup } from './FormGroup';
-export { default as Form } from './Form';
-export { default as Input } from './Input';
-export { default as Label } from './Label';
