+++ src/components/atoms/login/Content.js
@@ -1,5 +1,4 @@
 import styled from 'styled-components';
-import { breakpoint, size } from '../../base/breakpoints';
 
 const Content = styled.div`
   width: 320px;
@@ -8,12 +7,12 @@ const Content = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: center;
-  ${breakpoint(size.tablet)`
+  /* 
     width:450px;
-  `}
-  ${breakpoint(size.mobile)`
+ 
+  
     width:100%;
-  `}
+  */
 `;
 
 export default Content;
