--- src/components/atoms/typography/Title.js
@@ -1,15 +1,0 @@
-import styled from 'styled-components';
-
-import colors from '../../base/colors';
-
-const Title = styled.h2`
-  color: ${colors.darkBlue};
-  font-size: 34px;
-  font-weight: 300;
-  font-stretch: normal;
-  font-style: normal;
-  line-height: 0.94;
-  letter-spacing: normal;
-`;
-
-export default Title;
