--- src/components/atoms/typography/index.js
@@ -1,2 +1,0 @@
-export { default as Title } from './Title';
-export { default as Subtitle } from './Subtitle';
