--- src/components/base/colors.js
@@ -1,13 +1,0 @@
-export default {
-  EgyptianblueColor: ' #1e2a9c',
-  MidnightblueColor: '#15225a',
-  BrandeisblueColor: '#0273ff',
-  FlameColor: '#f25829',
-  GlitterColor: '#e0eaf8',
-  LightslategrayColor: '#797ca8',
-  gray: '#8798ad',
-  secondaryBlue: '#2e5bff',
-  secondaryGray: '#b0bac9',
-  darkBlue: '#2e384d',
-  whiteBackground: '#797ca8',
-};
