+++ src/components/molecules/appInput/style.scss
@@ -1,0 +1,31 @@
+.form-group {
+  margin: 10px 0;
+}
+
+.form-group input {
+  width: 100%;
+  height: 41px;
+  border-radius: 5px;
+  border: solid 1px #e0e7ff;
+  background-color: rgba(224, 231, 255, 0.2);
+  padding-left: 16px;
+  font-size: 15px;
+  font-weight: normal;
+  font-stretch: normal;
+  font-style: normal;
+  line-height: normal;
+  letter-spacing: normal;
+  color: #2e384d;
+}
+
+.form-group label {
+  width: 100%;
+  height: 14px;
+  font-size: 12px;
+  font-weight: 500;
+  font-stretch: normal;
+  font-style: normal;
+  line-height: normal;
+  letter-spacing: 1.13px;
+  /*  color: ${colors.secondaryGray}; */
+}
