+++ src/index.css
@@ -1,4 +1,4 @@
-body {
+/* body {
   margin: 0;
   padding: 0;
   box-sizing: border-box;
@@ -8,3 +8,4 @@ img {
   width: 100%;
   height: 100%;
 }
+ */
