+++ src/pages/login/index.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import './style.scss';
+
+/* import CardBase from '../../components/organisms/CardBase';
+import MainContainer from '../../components/atoms/main/MainContainer';
+import { ImageBackground } from '../../components/atoms/login'; */
+
+const Login = () => (
+  /*   <MainContainer flexDireccion='row'>
+    <CardBase />
+    <ImageBackground />
+  </MainContainer> */
+  <h1>hh</h1>
+);
+
+export default Login;
