+++ src/sass/base/mixins.scss
@@ -1,0 +1,14 @@
+$tablet: 768px;
+$mobile: 425px;
+
+@mixin tablet() {
+  @media only screen and (max-width: $tablet) {
+    @content;
+  }
+}
+
+@mixin mobile() {
+  @media only screen and (max-width: $mobile) {
+    @content;
+  }
+}
