+++ src/sass/main.scss
@@ -1,0 +1,9 @@
+@import './base/colors.scss';
+@import './base/mixins.scss';
+
+@import './global.scss';
+
+@import url('https://fonts.googleapis.com/css?family=Raleway:300,400,600');
+$font-family-sans-serif: 'Rubik', sans-serif;
+
+/* @import '../../../node_modules/bootstrap/scss/bootstrap.scss'; */
